% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set.}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{RIVER}{River the fish was found in}
  \item{MILE}{Mile along the river the fish was found in}
  \item{SPECIES}{The species of the fish}
  \item{LENGTH}{The length of the fish}
  \item{WEIGHT}{The weight of the fish}
  \item{DDT}{The DDT concentration in the fish}
  ...
}
}
\source{
Textbook
}
\usage{
ddt
}
\description{
A dataset containing the measurements of DDT levels in fish.
}
\keyword{datasets}
