% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firedam.R
\docType{data}
\name{firedam}
\alias{firedam}
\title{Fire damage data set.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance of property from damage, in miles}
  \item{DAMAGE}{damage done due to the fire, in thousands of dollars}
  ...
}
}
\source{
Textbook
}
\usage{
firedam
}
\description{
15 lines of measurements on property damage due to fire
}
\keyword{datasets}
