% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootleg function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations}

\item{x}{the created sample}

\item{fun}{the action run inside the function}

\item{alpha}{the percentage as a decimal of the data outside the confidence range}

\item{cx}{constant of x}

\item{...}{pulls parameters from within the function}
}
\value{
an invisible list
}
\description{
Bootleg function
}
\examples{

}
