% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Histogram of a binomial clt}
\usage{
mycltb(n, iter, p, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{p}{probability of a success}

\item{...}{take parameters from within the function}
}
\value{
a histogram
}
\description{
Histogram of a binomial clt
}
\examples{
\dontrun{mycltb(n=4,iter=10000,p=0.3)}
}
