% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson Function}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{the value of x at 1}

\item{delta}{used for a certain increment}

\item{llik}{a value of likelihood}

\item{xrange}{range of x values}

\item{parameter}{a way to not hard code the x label}
}
\value{
a log likelihood curve and Newton-Raphson Algorithm curve
}
\description{
Newton-Raphson Function
}
\examples{
\dontrun{myNRML(x0=1,delta=0.000001,llik=function(x) log(dpois(4,x)*dpois(6,x)*dpois(7,x)*dpois(6,x)*dpois(5,x)),xrange=c(0,20),parameter="lambda" )}
}
